postproc: Replaced inline asm for prefetching with prefetch functions
Prefetching functions are defined in postprocess_template using the
RENAME macro so that prefetching is used when available. For x86 targets
inline asm is used and the functions are non-empty only for cpus where
prefetching is available. For non x86 targets the gcc bultin prefetch is
used if it is available, otherwise no prefetching is done.