postproc: Moved inline asm for packing QP to seperate function
This patch contains the code for the avx2/sse2 versions of the new
function, but they are deliberately ignored, since the support for
avx2/sse2 isn't yet present (the next commit fixes this).

This is a temporary measure until full sse2/avx2 implementation is
complete, but it works with sse2/avx2 as inline asm.