postproc: Various formatting updates from the recent changes
I did my best to make as few changes as possible to the formatting when
adding new code, so this commit is just a means of making the format
changes that go along with the new code.

Mostly these are just changes in indentation, but I also re-formatted a
few assignment statments (from x= y -> x = y), and added some braces to
make some loops/if-else statements clearer.